Small renaming of files and moved DB classes into separate module
