+++ database/__init__.py
@@ -1,10 +1,7 @@
-import sys
-
 from argparse import ArgumentParser
 from configparser import ConfigParser
 
 from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
 
 __author__ = "Adam Bobowski"
@@ -19,17 +16,15 @@ def parse_args():
 args = parse_args()
 config = ConfigParser()
 config.read(args.config_file)
-print(config.sections())
 
-print([(a, b) for a, b in config["SQLite"].items()])
 
 class DB:
     Base = declarative_base()
     engine = create_engine(
         {
             "sqlite": "sqlite:///{path}",
-            "mysql": "{conn}://{user}:{pwd}@{url}/{db}"
-        }[config.sections()[0].tolower()].format(
+            "mysql": "{connection}://{login}:{password}@{url}/{database}"
+        }[config.sections()[0].lower()].format(
             **config[config.sections()[0]]
         )
     )
